OfficeClipboard: Re-connect each time
1. Use a local pipe instead of a TCP connection. We do not need the
extra overhead.

2. Reconnect each time. This is so that we can avoid a bug. Currently
whenever anyone tries to open a word document when LO is running (via
File -> Open) LO may crash. It crashes while trying to process an URP
request (Uno remote protocol). This script is what causes it. The script
also crashes with a "URP Bridge Disconnected" error.

There does not seem to be a convenient way to figure out if we can or
cannot make URP requests from the script. The only way seems to be to
try a fresh connection each time. It's a workaround, but it was either
this or that we increase the interval between our requests and hope for
the best. (Or fix LO)