Set davfs file_refresh cache to 5 seconds
This basically batches open operations during an amount of time. Again
really useful because libreoffice would open the same file many times
(with the same flags btw) resulting in multiple file downloads.

We want this to be a really low cache value because otherwise the app
will be desync from seafile resulting in a bad user experience.